+++ frontend/eventsync_front/src/types/questions.ts
